# backend/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SharedMoments API",
    "description": "API-Endpunkte für die SharedMoments API",
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
...
  "paths": {
    "/user": {
      "post": {
        "tags": ["Auth"],
        "summary": "Neuen Benutzer erstellen",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Benutzer erfolgreich erstellt"
          },
          "400": {
            "description": "Es sind nicht alle Werte gefüllt worden."
          },
          "500": {
            "description": "Fehler beim Datenbankzugriff"
          }
        }
      },
      "put": {
        "tags": ["Auth"],
        "summary": "Benutzerdaten aktualisieren",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Benutzer erfolgreich geändert"
          },
          "400": {
            "description": "Es sind nicht alle Werte gefüllt worden."
          },
          "500": {
            "description": "Fehler beim Datenbankzugriff"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "parameters": [
          {
            "name": "username",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Login erfolgreich"
          },
          "400": {
            "description": "Es sind nicht alle Werte gefüllt worden."
          },
          "500": {
            "description": "Fehler beim Datenbankzugriff"
          }
        }
      }
    },
    "/session": {
      "get": {
        "tags": ["Auth"],
        "summary": "Session überprüfen",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Session-ID ist gültig"
          },
          "401": {
            "description": "Session-ID ist ungültig oder abgelaufen"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Auth"],
        "summary": "Session beenden",
        "parameters": [
          {
            "name": "session_id",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Session erfolgreich beendet"
          },
          "400": {
            "description": "Es sind nicht alle Werte gefüllt worden."
          },
          "500": {
            "description": "Fehler beim Datenbankzugriff"
          }
        }
      }
    },
    "/setup": {
      "get": {
        "tags": ["Auth"],
        "summary": "Setup-Status überprüfen",
        "responses": {
          "200": {
            "description": "Setup ist abgeschlossen"
          },
          "500": {
            "description": "Fehler beim Datenbankzugriff"
          }
        }
      }
    },
...
  }
}